# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/search": {
      "get": {
        "operationId": "search_search",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "minLength": 2,
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "GAME",
                  "MOVIE",
                  "TV_SHOW",
                  "BOOK",
                  "ANIME"
                ]
              }
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "searchTerm",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TitleListResponse"
                }
              }
            }
          }
        },
        "tags": [
          "search"
        ]
      }
    },
...
    "/external/search": {
      "get": {
        "operationId": "integration_search",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "minLength": 2,
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "GAME",
                  "MOVIE",
                  "TV_SHOW",
                  "BOOK",
                  "ANIME"
                ]
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExternalTitleResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Поиск во внешних API — то, чего ещё нет в нашей базе",
        "tags": [
          "external"
        ]
      }
    },
    "/external/import": {
      "post": {
        "operationId": "integration_import",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportTitleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportedTitleResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Сохранить внешний тайтл к нам",
        "tags": [
          "external"
        ]
      }
    },
    "/external/quick-add": {
      "post": {
        "operationId": "integration_quickAdd",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickAddDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryEntryResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Добавление в один клик из расширения: импорт + библиотека",
        "tags": [
          "external"
        ]
      }
    },
...
    "/collections": {
      "get": {
        "operationId": "collection_findAll",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "searchTerm",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "collections"
        ]
      },
      "post": {
        "operationId": "collection_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "collections"
        ]
      }
    },
    "/collections/public/{username}/{slug}": {
      "get": {
        "operationId": "collection_findPublic",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicCollectionResponse"
                }
              }
            }
          }
        },
        "summary": "Публичная подборка по ссылке — без авторизации",
        "tags": [
          "collections"
        ]
      }
    },
    "/collections/{id}": {
      "get": {
        "operationId": "collection_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "collections"
        ]
      },
      "patch": {
        "operationId": "collection_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCollectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "collections"
        ]
      },
      "delete": {
        "operationId": "collection_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "collections"
        ]
      }
    },
    "/collections/{id}/titles": {
      "post": {
        "operationId": "collection_addTitles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionTitlesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "collections"
        ]
      },
      "delete": {
        "operationId": "collection_removeTitles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionTitlesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "collections"
        ]
      }
    },
    "/collections/{id}/share": {
      "post": {
        "operationId": "collection_share",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareCollectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Поделиться подборкой — получатели увидят уведомление",
        "tags": [
          "collections"
        ]
      }
    },
...
  },
  "info": {
    "title": "RED Marathōn API",
    "description": "One library for movies, TV shows, anime, books and games",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TitleListItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "enum": [
              "GAME",
              "MOVIE",
              "TV_SHOW",
              "BOOK",
              "ANIME"
            ],
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "coverUrl": {
            "type": "string",
            "nullable": true
          },
          "releaseDate": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "number"
          },
          "ratingCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "type",
          "name",
          "slug",
          "coverUrl",
          "releaseDate",
          "rating",
          "ratingCount"
        ]
      },
      "TitleListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TitleListItemResponse"
            }
          },
          "isHasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "isHasMore"
        ]
      },
...
      "ExternalTitleResponse": {
        "type": "object",
        "properties": {
          "externalId": {
            "type": "string"
          },
          "externalSource": {
            "type": "string",
            "enum": [
              "TMDB",
              "RAWG",
              "GOOGLE_BOOKS",
              "SHIKIMORI"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "GAME",
              "MOVIE",
              "TV_SHOW",
              "BOOK",
              "ANIME"
            ]
          },
          "name": {
            "type": "string"
          },
          "originalName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "coverUrl": {
            "type": "string"
          },
          "releaseDate": {
            "format": "date-time",
            "type": "string"
          },
          "rating": {
            "type": "number"
          },
          "ratingCount": {
            "type": "number"
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "externalId",
          "externalSource",
          "type",
          "name",
          "genres"
        ]
      },
      "ImportTitleDto": {
        "type": "object",
        "properties": {
          "source": {
            "type": "string",
            "enum": [
              "TMDB",
              "RAWG",
              "GOOGLE_BOOKS",
              "SHIKIMORI"
            ]
          },
          "externalId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "description": "TMDB отдаёт фильмы и сериалы по разным путям — тип нужен ему",
            "enum": [
              "GAME",
              "MOVIE",
              "TV_SHOW",
              "BOOK",
              "ANIME"
            ]
          }
        },
        "required": [
          "source",
          "externalId"
        ]
      },
      "ImportedTitleResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug"
        ]
      },
      "QuickAddDto": {
        "type": "object",
        "properties": {
          "source": {
            "enum": [
              "TMDB",
              "RAWG",
              "GOOGLE_BOOKS",
              "SHIKIMORI"
            ],
            "type": "string"
          },
          "externalId": {
            "type": "string"
          },
          "type": {
            "description": "TMDB отдаёт фильмы и сериалы по разным путям — тип нужен ему",
            "enum": [
              "GAME",
              "MOVIE",
              "TV_SHOW",
              "BOOK",
              "ANIME"
            ],
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PLANNED",
              "IN_PROGRESS",
              "COMPLETED",
              "ON_HOLD",
              "DROPPED"
            ]
          },
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 10
          }
        },
        "required": [
          "source",
          "externalId"
        ]
      },
      "LibraryEntryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "enum": [
              "PLANNED",
              "IN_PROGRESS",
              "COMPLETED",
              "ON_HOLD",
              "DROPPED"
            ],
            "type": "string"
          },
          "progress": {
            "type": "number",
            "nullable": true
          },
          "progressUnit": {
            "nullable": true,
            "enum": [
              "EPISODE",
              "CHAPTER",
              "PAGE",
              "HOUR",
              "PERCENT"
            ],
            "type": "string"
          },
          "rating": {
            "type": "number",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "isFavorite": {
            "type": "boolean"
          },
          "startedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "finishedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "title": {
            "$ref": "#/components/schemas/TitleListItemResponse"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "progress",
          "progressUnit",
          "rating",
          "note",
          "isFavorite",
          "startedAt",
          "finishedAt",
          "title",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CollectionListItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "coverUrl": {
            "type": "string",
            "nullable": true
          },
          "isPublic": {
            "type": "boolean"
          },
          "titlesCount": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "description",
          "coverUrl",
          "isPublic",
          "titlesCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "CollectionListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CollectionListItemResponse"
            }
          },
          "isHasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "isHasMore"
        ]
      },
      "CollectionOwnerResponse": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "username",
          "displayName",
          "avatarUrl"
        ]
      },
      "PublicCollectionResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "coverUrl": {
            "type": "string",
            "nullable": true
          },
          "titlesCount": {
            "type": "number"
          },
          "titles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TitleListItemResponse"
            }
          },
          "owner": {
            "$ref": "#/components/schemas/CollectionOwnerResponse"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "name",
          "slug",
          "description",
          "coverUrl",
          "titlesCount",
          "titles",
          "owner",
          "createdAt"
        ]
      },
      "CollectionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "coverUrl": {
            "type": "string",
            "nullable": true
          },
          "isPublic": {
            "type": "boolean"
          },
          "titlesCount": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "titles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TitleListItemResponse"
            }
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "description",
          "coverUrl",
          "isPublic",
          "titlesCount",
          "createdAt",
          "updatedAt",
          "titles"
        ]
      },
      "CreateCollectionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "coverUrl": {
            "type": "string",
            "maxLength": 500
          },
          "isPublic": {
            "type": "boolean"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCollectionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "coverUrl": {
            "type": "string",
            "maxLength": 500
          },
          "isPublic": {
            "type": "boolean"
          }
        }
      },
      "CollectionTitlesDto": {
        "type": "object",
        "properties": {
          "titleIds": {
            "minItems": 1,
            "maxItems": 50,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "titleIds"
        ]
      },
      "ShareCollectionDto": {
        "type": "object",
        "properties": {
          "usernames": {
            "minItems": 1,
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "usernames"
        ]
      },
...
    }
  }
}